git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@2044 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ magick/string-private.h
@@ -61,14 +61,14 @@ static inline int StringToInteger(const char *value)
   return((int) strtol(value,(char **) NULL,10));
 }
 
-static inline ssize_t StringToLong(const char *value)
+static inline long StringToLong(const char *value)
 {
-  return((ssize_t) strtol(value,(char **) NULL,10));
+  return(strtol(value,(char **) NULL,10));
 }
 
-static inline size_t StringToUnsignedLong(const char *value)
+static inline unsigned long StringToUnsignedLong(const char *value)
 {
-  return((size_t) strtoul(value,(char **) NULL,10));
+  return(strtoul(value,(char **) NULL,10));
 }
 
 #if defined(__cplusplus) || defined(c_plusplus)
